Changes Promise.run to run in order placed in queue
Currently whenever there is a promise waiting, they are pushed into PromiseMock.waiting, but run pops the last element and runs it synchronously. I want it to run the promises in the order they were placed so I am changing it to use shift in the run function. Had to change all tests that used Promise.reject to call Promise.run(2) because Promise.reject returns a promise object.